moved test funcs
