+++ pulsar-broker/src/test/java/org/apache/pulsar/broker/loadbalance/extensions/ExtensibleLoadManagerImplBaseTest.java
@@ -20,14 +20,9 @@ package org.apache.pulsar.broker.loadbalance.extensions;
 
 import static org.mockito.Mockito.reset;
 import static org.mockito.Mockito.spy;
-import static org.testng.Assert.assertFalse;
-import static org.testng.Assert.assertTrue;
 import com.google.common.collect.Sets;
-
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeUnit;
 import org.apache.commons.lang3.reflect.FieldUtils;
 import org.apache.commons.lang3.tuple.Pair;
 import org.apache.pulsar.broker.PulsarService;
@@ -37,8 +32,6 @@ import org.apache.pulsar.broker.loadbalance.extensions.channel.ServiceUnitStateC
 import org.apache.pulsar.broker.loadbalance.extensions.scheduler.TransferShedder;
 import org.apache.pulsar.broker.testcontext.PulsarTestContext;
 import org.apache.pulsar.client.admin.PulsarAdminException;
-import org.apache.pulsar.client.api.PulsarClient;
-import org.apache.pulsar.client.api.PulsarClientException;
 import org.apache.pulsar.client.impl.LookupService;
 import org.apache.pulsar.common.naming.NamespaceBundle;
 import org.apache.pulsar.common.naming.SystemTopicNames;
@@ -171,37 +164,5 @@ public abstract class ExtensibleLoadManagerImplBaseTest extends MockedPulsarServ
             }
             i++;
         }
-    }
-
-    protected static LookupService spyLookupService(PulsarClient client) throws IllegalAccessException {
-        LookupService svc = (LookupService) FieldUtils.readDeclaredField(client, "lookup", true);
-        var lookup = spy(svc);
-        FieldUtils.writeDeclaredField(client, "lookup", lookup, true);
-        return lookup;
-    }
-
-    protected static void checkOwnershipState(String broker, NamespaceBundle bundle,
-                                              ExtensibleLoadManager primaryLoadManager,
-                                              ExtensibleLoadManager secondaryLoadManager, PulsarService pulsar1)
-            throws ExecutionException, InterruptedException {
-        var targetLoadManager = secondaryLoadManager;
-        var otherLoadManager = primaryLoadManager;
-        if (broker.equals(pulsar1.getBrokerServiceUrl())) {
-            targetLoadManager = primaryLoadManager;
-            otherLoadManager = secondaryLoadManager;
-        }
-        assertTrue(targetLoadManager.checkOwnershipAsync(Optional.empty(), bundle).get());
-        assertFalse(otherLoadManager.checkOwnershipAsync(Optional.empty(), bundle).get());
-    }
-    protected void checkOwnershipState(String broker, NamespaceBundle bundle)
-            throws ExecutionException, InterruptedException {
-        checkOwnershipState(broker, bundle, primaryLoadManager, secondaryLoadManager, pulsar1);
-    }
-
-    protected static PulsarClient pulsarClient(String url, int intervalInSecs) throws PulsarClientException {
-        return
-                PulsarClient.builder()
-                        .serviceUrl(url)
-                        .statsInterval(intervalInSecs, TimeUnit.SECONDS).build();
     }
 }
